setup dassets file and assets
This allows for package management using Dassets.  This assumes
you will be using these combinatins in a larger combination that
includes Romo's base/modal/dropdown components or that that component
scripts will be loaded prior to loading these scripts.

This is part of setting up romo-av to work with romo and dassets
as a separate component.