video js: more cleanups to video unit calculations
This does some updates to how video unit calculations are done to
try and normalize frame number handling and errors due to limitations
in how precise video objects specify their duration.

The problem is that video that is `1.167`s long and has a  `24`fps
frame rate ends up having a frame count of `28.0080000003` frames.
This doesn't make any sense (it should only have `28` frames).  If
you assume `28` frames and calculate the duration from the frame
rate you get `1.166666666667`.  So the browser is limiting the
precision of the length to the millisecond and rounding any remainder.

To account for this limitation, we switch to calculating the duration
frames by rounding any value we calculate.  This is based on the
assumption that video codecs do not allow for partial frames so
any partial frame in our calculations can rounded away as precision
limitation.  This also switches to hard-coding the duraiton percent
to `100` as there is no need to calculate this value.

Now that we have a proper duration frame count, we need to account
for converting playback and arbitrary times to frame counts.  To
do this, we switch to first computing the percentage of the given
time and then converting that percentage to frame count.  This moves
the frame calculation to a scale that is relative to the duration
time (even if it has been inaccurately rounded).

This also switches to calculating frames from percent to use `ceil`
for positive numbers and `floor` for negative numbers.  The effect
is that if a given percent lands right on a frame boundary the
number closest to zero will be chosen while still preserving the
behavior that partial frame calculations will be rounded up.  The
goal is that if given 100%, the frame calc will land on the boundary
ending the last frame and will be counted as the last frame's number.
A side-effect is that zero percent is numbered as the zero frame.
This is technically accurate in that if you haven't started the movie
you can't be on frame `1`.

Some other cleanups included here:

* cache the non-static duration units and update on new meta data as
  there is no need to re-calculate these over and over again
* more thoroughly detect invalid (`NaN`) durations