update video js to use just vanilla js and Romo's new API
This is part of updating Romo to not require jquery and to use a
new vanilla js api.  This switches to the new Romo equivalents
where appropriate.  Since most of this was already written in a
vanilla js friendly manner (b/c we were using the native video obj
anyway), this mostly means updating event handling and property
handling.