update modal video js to use Romo's new API, vanilla js
This is part of porting Romo to not depend on jQuery anymore.  We
are moving Romo to a vanilla js implementation and this covers the
modal video component.  This mostly consists of event handling
updates for the proxy events and updating to init the modal/video
components.  This also switches to Romo's new auto init API.