Update video, dropdown video and modal video to use `RomoComponent`
This updates the video, dropdown video, and modal video components
to use the `RomoComponent` function to define their classes and
switches to the new init UI logic. This is part of updating to
work with the latest changes from Romo.

This includes updating the video component to define use the
new event function logic provided by `RomoComponent` and also
removes the empty `doInit` functions as the `RomoComponent`
helper provides these automatically.